Refactor InitFile into new module
The motivation here is to introduce a new module for containing all code
related to forged alliance lua, that is required for lobby use.

Should these become generic enough, they should be migrated to
faftools.lua.
