+++ src/fa/lua/__init__.py
@@ -1,0 +1,1 @@
+from .init_file import InitFile
