+++ tests/fa/test_integration.py
@@ -1,10 +1,7 @@
 __author__ = 'Sheeo'
 
-from fa.init_file import InitFile
 from fa.mod import Mod
-
 from git import Version
-
 from config import Settings
 
 """
